requirements.txt
python-telegram-bot==20.5
gspread==5.8.0
oauth2client==4.1.3